# cuDNN Flash Attention Forward & Backwards BF16 (+35% performance)