--- a/profile_gpt2.cu
+++ b/profile_gpt2.cu
@@ -50,6 +50,10 @@ int main() {
     // setup the (global) cuBLASLt workspace
     cudaCheck(cudaMalloc(&cublaslt_workspace, cublaslt_workspace_size));
 
+    #ifdef ENABLE_CUDNN
+    checkCudnnErr(cudnnCreate(&cudnn_handle));
+    #endif
+
     // build the GPT-2 model from a checkpoint
     GPT2 model;
     gpt2_build_from_checkpoint(&model, "gpt2_124M_bf16.bin");
@@ -76,6 +80,12 @@ int main() {
     cudaCheck(cudaDeviceSynchronize()); // finish all CUDA work to get correct precise timings
     // free
     gpt2_free(&model);
+
+    #ifdef ENABLE_CUDNN
+    if (cudnn_workspace != NULL) {
+        cudaCheck(cudaFree(cudnn_workspace));
+    }
+    #endif
     cudaCheck(cudaFree(cublaslt_workspace));
     cublasCheck(cublasDestroy(cublas_handle));
     cublasCheck(cublasLtDestroy(cublaslt_handle));