--- a/profile_gpt2.cu
+++ b/profile_gpt2.cu
@@ -86,6 +86,7 @@ int main() {
         cudaCheck(cudaFree(cudnn_workspace));
     }
     #endif
+    checkCudnnErr(cudnnDestroy(cudnn_handle));
     cudaCheck(cudaFree(cublaslt_workspace));
     cublasCheck(cublasDestroy(cublas_handle));
     cublasCheck(cublasLtDestroy(cublaslt_handle));